[HACK] Changes token to i128 bit
Cassandra clusters using RandomPartitioner generate 128-bit tokens.
This is a hack to fix a parsing error when connecting to these clusters.
Note that the Token Aware load balancer still ASSUMES Murmur3 and is
will not load balance correctly if the Cassandra OR Scylla cluster are
using a different load balancing strategy.